Some py2 to 3 fixes. Moving keystone/swift to separate directory
